[feat/14] refactor: 그리드/구독, 전체/그리드, 구독/리스트 렌더링 추출, 상황별 이벤트 모임 분리 중
