+++ README.md
@@ -14,3 +14,5 @@ FE 1,2 주 프로젝트
 - feat/#10 : snack bar 구현
 - feat/#11 : grid 해지하기 버튼 구현
 - feat/#12 : grid 구독하기 버튼 구현
+- feat/#13 : 구독, 해지 후속 동작 추가
+- feat/#14 : 렌더 로직, 이벤트 핸들러 리팩토링
