+++ components/articleList/event/eventHandlers.js
@@ -1,0 +1,10 @@
+import { addCategorySelectionEventListener,
+    addScrollEventListener,
+    addModeSelectionEventListener,
+} from "./clickEvent.js"
+
+export const handleGridSubscription = () => {
+    addCategorySelectionEventListener();
+    addScrollEventListener();
+    addModeSelectionEventListener();
+}
